Move string-prefix enumerations to a separate submodule
