+++ crates/ruff_python_ast/src/lib.rs
@@ -21,6 +21,7 @@ pub mod relocate;
 pub mod statement_visitor;
 pub mod stmt_if;
 pub mod str;
+pub mod str_prefix;
 pub mod traversal;
 pub mod types;
 pub mod visitor;
